Move update and file related things to data.py
